requirements.txt
aniso8601==8.0.0
attrs==20.1.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
flake8==3.8.3
Flask==1.1.2
flask-restplus==0.13.0
idna==2.10
inflection==0.5.1
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.5.0
mypy==0.782
mypy-extensions==0.4.3
openapi-spec-validator==0.2.9
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==6.0.1
pytest-flask==1.0.0
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rope==0.17.0
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==0.16.1
